Conflicts:
	Babylon/Materials/textures/babylon.texture.js
	Babylon/Materials/textures/babylon.texture.ts
	Babylon/babylon.engine.js
	Babylon/babylon.engine.ts

+++ Babylon/Shaders/color.fragment.fx
@@ -1,4 +1,4 @@
-﻿precision mediump float;
+﻿precision highp float;
 
 uniform vec4 color;
